+++ src/ArduinoJson/Namespace.hpp
@@ -10,16 +10,16 @@
 
 #ifndef ARDUINOJSON_VERSION_NAMESPACE
 
-#  define ARDUINOJSON_VERSION_NAMESPACE                                    \
-    ARDUINOJSON_CONCAT4(ARDUINOJSON_VERSION_MACRO,                         \
-                        ARDUINOJSON_BIN2ALPHA(ARDUINOJSON_ENABLE_PROGMEM,  \
-                                              ARDUINOJSON_USE_LONG_LONG,   \
-                                              ARDUINOJSON_USE_DOUBLE, 1),  \
-                        ARDUINOJSON_BIN2ALPHA(ARDUINOJSON_ENABLE_NAN,      \
-                                              ARDUINOJSON_ENABLE_INFINITY, \
-                                              ARDUINOJSON_ENABLE_COMMENTS, \
-                                              ARDUINOJSON_DECODE_UNICODE), \
-                        ARDUINOJSON_SLOT_ID_SIZE)
+#  define ARDUINOJSON_VERSION_NAMESPACE                               \
+    ARDUINOJSON_CONCAT5(                                              \
+        ARDUINOJSON_VERSION_MACRO,                                    \
+        ARDUINOJSON_BIN2ALPHA(ARDUINOJSON_ENABLE_PROGMEM,             \
+                              ARDUINOJSON_USE_LONG_LONG,              \
+                              ARDUINOJSON_USE_DOUBLE, 1),             \
+        ARDUINOJSON_BIN2ALPHA(                                        \
+            ARDUINOJSON_ENABLE_NAN, ARDUINOJSON_ENABLE_INFINITY,      \
+            ARDUINOJSON_ENABLE_COMMENTS, ARDUINOJSON_DECODE_UNICODE), \
+        ARDUINOJSON_SLOT_ID_SIZE, ARDUINOJSON_STRING_LENGTH_SIZE)
 
 #endif